refactoring: change plugin name.
